Grafana: Add make target for local dev
When modifying dashboards it can be useful to have a live Grafana
instance with access to a Prometheus datasource with metric data.

We can run a local Grafana with Docker and thanks to Grafana's
`provisioning` support we can have it watch for dashboard changes after
running `make` and reload them.

Run `make grafana` and start iterating...