Gitlab-CI: Correct asset naming
By default Gitlab will provide source in Release with
`spinnaker-mixin.tar.gz` so we need to avoid clashing with that
filename. Add `-rendered` to denote that it is rendered manifests not
raw source code (jsonnet/etc)

Also, we should have the TAG in the rendered tar.gz so it can be
differentiated when downloaded.